Move pages into chapter folders
