+++ src/applications/pensions/config/chapters/health-and-employment-information/medicalCondition.js
@@ -2,29 +2,29 @@ import {
   yesNoSchema,
   yesNoUI,
 } from '@department-of-veterans-affairs/platform-forms-system/web-component-patterns';
-import { MedicalEvidenceAlert } from '../components/FormAlerts';
+import { MedicalEvidenceAlert } from '../../../components/FormAlerts';
 
 /** @type {PageSchema} */
 export default {
   uiSchema: {
-    'ui:title': 'Social Security disability',
-    socialSecurityDisability: yesNoUI({
-      title: 'Do you currently receive Social Security disability payments?',
+    'ui:title': 'Medical condition',
+    medicalCondition: yesNoUI({
+      title: 'Do you have a medical condition that prevents you from working?',
       uswds: true,
       classNames: 'vads-u-margin-bottom--2',
     }),
     'view:warningAlert': {
       'ui:description': MedicalEvidenceAlert,
       'ui:options': {
-        hideIf: formData => formData.socialSecurityDisability !== false,
+        hideIf: formData => formData.medicalCondition !== true,
       },
     },
   },
   schema: {
     type: 'object',
-    required: ['socialSecurityDisability'],
+    required: ['medicalCondition'],
     properties: {
-      socialSecurityDisability: yesNoSchema,
+      medicalCondition: yesNoSchema,
       'view:warningAlert': {
         type: 'object',
         properties: {},
