+++ src/applications/pensions/config/chapters/health-and-employment-information/vaTreatmentHistory.js
