+++ src/applications/pensions/config/chapters/military-history/generalHistory.js
