+++ src/applications/pensions/config/form.js
@@ -55,28 +55,28 @@ import monthlyIncomeUI from '../definitions/monthlyIncome';
 import expectedIncomeUI from '../definitions/expectedIncome';
 import { additionalSourcesSchema } from '../definitions/additionalSources';
 import otherExpensesUI from '../definitions/otherExpenses';
-import applicantInformation from '../pages/applicantInformation';
-import mailingAddress from '../pages/mailingAddress';
-import contactInformation from '../pages/contactInformation';
-import servicePeriods from '../pages/servicePeriods';
-import generalHistory from '../pages/generalHistory';
-import pow from '../pages/pow';
-import age from '../pages/age';
-import socialSecurityDisability from '../pages/socialSecurityDisability';
-import medicaidCoverage from '../pages/medicaidCoverage';
-import medicaidStatus from '../pages/medicaidStatus';
-import medicalCondition from '../pages/medicalCondition';
-import nursingHome from '../pages/nursingHome';
-import specialMonthlyPension from '../pages/specialMonthlyPension';
-import vaTreatmentHistory from '../pages/vaTreatmentHistory';
-import federalTreatmentHistory from '../pages/federalTreatmentHistory';
-import generateMedicalCentersSchemas from '../pages/medicalCenters';
-import currentEmployment from '../pages/currentEmployment';
-import generateEmployersSchemas from '../pages/employmentHistory';
-import maritalStatus from '../pages/maritalStatus';
-import currentSpouse from '../pages/currentSpouse';
-import dateOfCurrentMarriage from '../pages/dateOfCurrentMarriage';
-import reasonForCurrentSeparation from '../pages/reasonForCurrentSeparation';
+import applicantInformation from './chapters/applicant-information/applicantInformation';
+import mailingAddress from './chapters/applicant-information/mailingAddress';
+import contactInformation from './chapters/applicant-information/contactInformation';
+import servicePeriods from './chapters/military-history/servicePeriods';
+import generalHistory from './chapters/military-history/generalHistory';
+import pow from './chapters/military-history/pow';
+import age from './chapters/health-and-employment-information/age';
+import socialSecurityDisability from './chapters/health-and-employment-information/socialSecurityDisability';
+import medicaidCoverage from './chapters/health-and-employment-information/medicaidCoverage';
+import medicaidStatus from './chapters/health-and-employment-information/medicaidStatus';
+import medicalCondition from './chapters/health-and-employment-information/medicalCondition';
+import nursingHome from './chapters/health-and-employment-information/nursingHome';
+import specialMonthlyPension from './chapters/health-and-employment-information/specialMonthlyPension';
+import vaTreatmentHistory from './chapters/health-and-employment-information/vaTreatmentHistory';
+import federalTreatmentHistory from './chapters/health-and-employment-information/federalTreatmentHistory';
+import generateMedicalCentersSchemas from './chapters/health-and-employment-information/medicalCenters';
+import currentEmployment from './chapters/health-and-employment-information/currentEmployment';
+import generateEmployersSchemas from './chapters/health-and-employment-information/employmentHistory';
+import maritalStatus from './chapters/household-information/maritalStatus';
+import currentSpouse from './chapters/household-information/currentSpouse';
+import dateOfCurrentMarriage from './chapters/household-information/dateOfCurrentMarriage';
+import reasonForCurrentSeparation from './chapters/household-information/reasonForCurrentSeparation';
 
 import { validateAfterMarriageDate } from '../validation';
 import migrations from '../migrations';
