+++ src/applications/pensions/tests/chapters/health-and-employment-information/age.unit.spec.jsx
@@ -1,14 +1,14 @@
 import {
   testNumberOfErrorsOnSubmitForWebComponents,
   testNumberOfWebComponentFields,
-} from './pageTests.spec';
-import formConfig from '../../config/form';
-import medicaidCoverage from '../../pages/medicaidCoverage';
+} from '../pageTests.spec';
+import formConfig from '../../../config/form';
+import age from '../../../config/chapters/health-and-employment-information/age';
 
-const { schema, uiSchema } = medicaidCoverage;
+const { schema, uiSchema } = age;
 
-describe('medicaid coverage pension page', () => {
-  const pageTitle = 'medicaid coverage';
+describe('pension age page', () => {
+  const pageTitle = 'Age';
   const expectedNumberOfFields = 1;
   testNumberOfWebComponentFields(
     formConfig,
@@ -18,7 +18,7 @@ describe('medicaid coverage pension page', () => {
     pageTitle,
   );
 
-  const expectedNumberOfErrors = 0;
+  const expectedNumberOfErrors = 1;
   testNumberOfErrorsOnSubmitForWebComponents(
     formConfig,
     schema,
