+++ src/applications/pensions/tests/chapters/household-information/currentSpouse.unit.spec.jsx
@@ -1,9 +1,9 @@
 import {
   testNumberOfErrorsOnSubmitForWebComponents,
   testNumberOfWebComponentFields,
-} from './pageTests.spec';
-import formConfig from '../../config/form';
-import currentSpouse from '../../pages/currentSpouse';
+} from '../pageTests.spec';
+import formConfig from '../../../config/form';
+import currentSpouse from '../../../config/chapters/household-information/currentSpouse';
 
 describe('current spouse page', () => {
   it('should render web components in form correctly', () => {
