+++ src/applications/pensions/tests/chapters/military-history/pow.unit.spec.jsx
@@ -11,22 +11,22 @@ import {
 } from '@department-of-veterans-affairs/platform-forms-system/ui';
 
 import { DefinitionTester } from '@department-of-veterans-affairs/platform-testing/schemaform-utils';
-import getData from '../fixtures/mocks/mockStore';
+import getData from '../../fixtures/mocks/mockStore';
 
 import {
   testNumberOfErrorsOnSubmitForWebComponents,
   testNumberOfWebComponentFields,
-} from './pageTests.spec';
-import formConfig from '../../config/form';
-import generalHistory from '../../pages/generalHistory';
+} from '../pageTests.spec';
+import formConfig from '../../../config/form';
+import pow from '../../../config/chapters/military-history/pow';
 
 const definitions = formConfig.defaultDefinitions;
 
-const { schema, uiSchema } = generalHistory;
+const { schema, uiSchema } = pow;
 
 describe('web component tests', () => {
   const pageTitle = 'military history';
-  const expectedNumberOfFields = 2;
+  const expectedNumberOfFields = 1;
   testNumberOfWebComponentFields(
     formConfig,
     schema,
@@ -45,10 +45,10 @@ describe('web component tests', () => {
   );
 });
 
-describe('pension applicant information page', () => {
+describe('pension pow information page', () => {
   const middleware = [];
   const mockStore = configureStore(middleware);
-  it('should submit with no errors when No is selected', async () => {
+  it('should submit with valid data', async () => {
     const onSubmit = sinon.spy();
     const { data } = getData({ loggedIn: false });
     const { queryByText, container } = render(
@@ -74,10 +74,10 @@ describe('pension applicant information page', () => {
       expect(onSubmit.called).to.be.true;
     });
   });
-  it('should not submit with errors when Yes is selected', async () => {
+  it('should reveal date fields', async () => {
     const onSubmit = sinon.spy();
     const { data } = getData({ loggedIn: false });
-    const { queryByText, container } = render(
+    const { container } = render(
       <Provider store={mockStore(data)}>
         <DefinitionTester
           definitions={definitions}
@@ -89,47 +89,19 @@ describe('pension applicant information page', () => {
         />
       </Provider>,
     );
-    const submitBtn = queryByText('Submit');
-    const changeEvent = new CustomEvent('selected', {
-      detail: { value: 'Y' },
-    });
-    $('va-radio', container).__events.vaValueChange(changeEvent);
 
-    fireEvent.click(submitBtn);
+    // Verify fields not visible
+    expect($$('input,select', container).length).to.equal(0);
+
     await waitFor(() => {
-      expect(onSubmit.called).to.be.false;
-    });
-  });
-  it('should reveal name fields', async () => {
-    const onSubmit = sinon.spy();
-    const { data } = getData({ loggedIn: false });
-    const store = mockStore(data);
-    const { container } = render(
-      <Provider store={store}>
-        <DefinitionTester
-          definitions={definitions}
-          schema={schema}
-          uiSchema={uiSchema}
-          data={{}}
-          formData={{}}
-          onSubmit={onSubmit}
-        />
-      </Provider>,
-    );
-
-    waitFor(() => {
-      expect($$('input', container).length).to.equal(1);
-      expect($$('va-radio', container).length).to.equal(1);
-      expect($('button[type="submit"]', container)).to.exist;
-
       const changeEvent = new CustomEvent('selected', {
         detail: { value: 'Y' },
       });
       $('va-radio', container).__events.vaValueChange(changeEvent);
 
-      expect($$('va-radio-option', container).length).to.eq(2);
       // Verify fields are now visible
-      expect($$('input', container).length).to.equal(4);
+      expect($$('select', container).length).to.equal(4);
+      expect($$('input', container).length).to.equal(2);
     });
   });
 });
